backend/requirements/production.txt
# Production dependencies
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
django-extensions==3.2.3
django-ratelimit==4.1.0
django-storages==1.14.2
django-environ==0.11.2

# Database
psycopg2-binary==2.9.7
redis==5.0.1
hiredis==2.2.3

# Authentication & Security
djangorestframework-simplejwt==5.3.0
cryptography==41.0.8
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Background tasks
celery[redis]==5.3.4
flower==2.0.1

# API & Serialization
drf-spectacular==0.27.0
drf-spectacular-sidecar==0.1.5
marshmallow==3.20.1
marshmallow-enum==1.5.1

# File processing
Pillow==10.1.0
python-magic==0.4.27
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
openpyxl==3.1.2

# External integrations
requests==2.31.0
aiohttp==3.9.1
httpx==0.25.2
boto3==1.34.0
botocore==1.34.0

# Data processing
pandas==2.1.4
numpy==1.25.2
openpyxl==3.1.2
xlsxwriter==3.1.9

# Validation & Serialization
pydantic==2.5.0
email-validator==2.1.0
phonenumbers==8.13.25
python-dateutil==2.8.2

# Monitoring & Logging
sentry-sdk==1.38.0
structlog==23.2.0
prometheus-client==0.19.0

# Performance
django-debug-toolbar==4.2.0
django-silk==5.0.4

# Utilities
python-decouple==3.8
click==8.1.7
rich==13.7.0
typer==0.9.0

# Production server
gunicorn==21.2.0
uvicorn==0.24.0

# Security
bandit==1.7.5
safety==2.3.5